Fix compose file configuration for node-validator and block-explorer
Fix docker-compose and process-compose for node-validator

The `node-validator` section of the `docker-compose.yaml` file has the environment variable
`ESPRESSO_NODE_VALIDATOR_INITIAL_NODE_PUBLIC_BASE_URLS` pointing at the incorrect
ports for all of the sequencers that are not `sequencer0`.  As a result, the other sequencers are
not correctly scraped and are not reflected in the data collected.  Additionally, the port that the
`node-validator` is meant to expose is not being forwarded to the container via the environment
variable `ESPRESSO_NODE_VALIDATOR_PORT`.  Finally, the local demo service doesn't produce
any visible logs while running.

Forwards the relevant needed environment variables for logging, and modifies the the ports in
the node public base URLs to correct the issue.

Add node validator configuration to process-compose and docker-compose

The block-explorer now has support for the node validator API, but has not been updated to point
to the relevant base URL for the service.  As a result, the Nodes Page will still generate fake data.

Updates the process-compose.yaml and docker-compose.yaml configuration to supply the
environment variable `NODE_VALIDATOR_URI` so that it points to the `node-validator` as a
websocket connection.