Add tests for ProxyGlobalState::claim_block_header_input
Now that we have separated the implementation of `claim_block_header_input` from
the API declaration, so that we can have it return the specific error enum
variant `ClaimBlockHeaderInputError`, we have the ability to add unit tests to
trigger its specific error scenarios.

This adds unit tests to cover teh various error cases.

Additionally, comments have been added to various structus and methods to help
further the understanding of what these do, what they are for, and to draw
attention as to whether they need to stay around, or if their purpose needs to
be rethought.