Add unit tests for GlobalState::register_builder_state
These unit tests verify that the state modifications that do happen as a result of
invoking `register_builder_state` are expected, and indicate the current behavior.
Additionally, they serve a purpose to draw attention to these behaviors in an
effort to verify that they are the intended correct behavior, and not some missing
edge case.