Add unit tests for `handle_da_event`
At the moment `handle_da_event` doesn't have a return value which makes it difficult to
create unit tests that target specific behavior regarding successes or failures.  In order to
effectively test this function, an implementation function has been created for it called
`handle_da_event_implementation` which does return a new error enum type,
`HandleQcEventError` that encodes all of the various failure conditions that apply when
attempting to handle the qc event.  The code to `handle_qc_event` is then just updated
to invoke this `handle_qc_event_implementation` function, and to ignore the error in
order to maintain the previous function signature and behavior.

In addition to this refactoring, the processing of `handle_qc_event_implementation` has
been modified to take the happy path coding paradigm of early returns in the case of
errors.

With the creation of the `handle_qc_event_implementation` function now in existence the
creation of unit tests that target the specific error conditions can now be created in order
to trigger this conditions explicitly, as well as the success cases.  Unit tests that address
these conditions explicitly have been added.