Remove cached block heights
With the introduction of transaction isolation, concurrent writers
and readers, possibly working off different snapshots of the database,
it has become increasingly difficult to keep the in-memory cached heights
synchronized with the database. We run into the following trilemma:

* Consistency: the in-memory cache is always synchronized with the database,
  and never allows a reader to see a different value than they would if they
  read the heights from the database itself
* Performance: a reader never blocks on a writer
* Safety: no possibility for deadlocks

Solving this trilemma essentially amounts to implementing an ACID database,
so it no longer provides much benefit over just letting the databsae synchronize
access to heights.

Since we now allow concurrent connections to the database, access to the
database is no longer such a bottleneck, and the benefits of caching these
values no longer justify their added complexity. Moreover, the only accesses
to the heights on the critical path were checks to fail fast when data is not
available, but the case where the data is unavailable is exceptional and rare,
so optimizing it doesn't matter much anyways. Thus, this change also completely
removes two of the most impactful accesses to the cached heights.