Fix missed notification race condition
Previously, we were sending notifications for newly inserted objects
while those inserts were still isolated to an uncommitted transaction.
This lead to a situation where a client could subscribe to notifications
for an object just after the notification for that object was sent, but
before the object was committed to storage. Then the client would check
storage, conclude that the object was missing, and go to sleep, assuming
it would get a notification when the object arrived.

The fix is to only send notifications after the corresponding objects
have already been committed to storage.