Update query service, implement transaction isolation
This commit moves us to the new transaction-oriented version of the
query service. By moving most operations away from the main data source
into isolated transactions, we massively improve concurrency and eliminate
bottlenecks due to exclusive locks on the API state.

This also prepares us to switch to the sqlx backend and then replace the
file system persistence with sqlite.