Avoid blocking thread in transaction drop handler
When reverting a transaction on drop, we should not block on completion
of the revert future, because we may be blocking an executor thread,
potentially causing a deadlock. Instead, we simply queue the revert, so
that it is not actually executed until the next time the connection is
used in an async context (e.g. just before acquiring the connection for
a subsequent transaction). This is how sqlx handles the same problem.