Add `BuiltFromProposedBlock` to `spawned_builder_states`
In order to be able to consistently pick the correct `BuilderState` to extend when
we receive a `QuorumProposal`, additional information is needed when understanding
the spawned `BuilderState`s. In order to add this additional information the
`BuiltFromProposedBlock` struct has been added as a `Pair` to the value stored
in the `spawned_builder_states`.

This addition will allow for inspectors to better understand the `QuorumProposal`
information that the `BuilderState` was spawned from.  This field is also made an
`Option` to make it easier to initialize.