Add some-test.rs integration test
Adds `tests` crate w/ a single integration test. This test
re-implements `demo-smoke-test` bash script in rust, with some minor
additions. In order to avoid code duplication, some logic from
`espresso-bridge` is factored into a `client` crate, and imported from
there for use in the test. `espresso-bridge` is also updated to use
this `client`.

This primary goal of this PR is parity with `smoke-test` bash
script. But, there are a couple of improvements:

  * the number of blocks to wait before declaring the test successful
  is now configurable. This may be useful in scenarios like weekly
  hostshot integrations.
  * separate waits for for service readiness has been parallelized
  * adds a justfile entry