Refactor `process_quorum_proposal` method to "Happy Path"
The `process_quorum_proposal` method currently has a few nested `if` statements that
do some basic logic and exit.  These have been modified to move the else conditions
further up in the code in order to favor an early-return so as to keep the "Happy Path"
out of a conditional path.

A comment about what `process_quorum_proposal` does has also been added to help
add context for what this method is doing, and the conditions under which it operates.