Fix node-validator's state becoming stale due to disconnected leaf stream
The Leaf Stream that is sourced from the HotShot Query Service can becomes disconnected
due to a variety of issues that are not accounted for within the node-validator API.
These reasons can be varied, but they all have the same result, the failure to retrieve
a Stream, or the end of an existing Stream.

In either of these cases the node-validator will not attempt to re-establish the Leaf
Stream, resulting in the validator's state becoming stale over time.

In order to address this issue effectively, and ideally, changes have been made so
attempts to re-acquire the initial Leaf Stream state.  If this retrieval fails then
it will attempt again and again with an exponential backoff.  If the failure persists
for too long, the process will eventually give up and panic instead.