Add Option values to Explorer histogram data
The data in the Explorer histogram can occasionally have a missing component.  There
are times when we have written Block Header information without having written the
Block Payload. Due to this, it allows for the possibility of being able to query
data for a Block that has the Header, but no paylaod.  This will ultimately result
in an error as the size of the block in question can not be determined.  This will
currently result in an error.

The change is to allow for nullable values for things like block size.  It should not
happen with block time, but since block time is also a Option result, it doesn't hurt
to also allow for it's value to be an Option as well.