Add logic to populate initial node identity from stake table (#2099)
It was noticed that the initial list of node identities does not consider the public
keys listed in the stake table.  This can easily be demonstrated by launching the
`node-metrics` binary with `ESPRESSO_NODE_VALIDATOR_STAKE_TABLE_SOURCE_BASE_URL`
and `ESPRESSO_NODE_VALIDATOR_LEAF_STREAM_SOURCE_BASE_URL` populated, but missing
`ESPRESSO_NODE_VALIDATOR_INITIAL_NODE_PUBLIC_BASE_URLS`.

In this case the list of node identities will be empty, and the information coming down
to consuming clients will be missing all node identities.

The fix is to modify the `DataState` structure so that it is initially populated with
information from the passed in stake table.