Add behavior to allow for a grace period of empty blocks
Under normal circumstances the Builder never wants to return an empty
block.  However, since it takes a few rounds for consensus to finalize
transactions submitted to HotShot, it is a beneficial UX to make an
effort to finalize any submitted transactions as quickly as possible.

To that end, this pull request adds an option to the legacy builder to
allow for a grace period of empty block proposal after receiving a
message from consensus that indicates that a decide was not missing
transactions.  When in this grace period, the Builder will propose empty
blocks, if no other transactions are able to be submitted, up until the
grace period has ended.