Refactor notifications to simplify liveness edge cases
The previous commit fixed a subtle deadlock due to a missed notification
for a fetched object. This deadlock was itself caused by an imperfect
solution to a previously discovered missed notification bug. Clearly,
this area of the system is subtle and fraught.

The fix in the previous commit was itself imperfect:
* Because notifications were sent when a method on `notify_storage::Transaction`
  is called (e.g. `insert_leaf`) this solution did not and could not handle the
  case where storage fails to even open a transaction. This commit adds a test
  `test_fetch_storage_failure_on_begin` which fails on the previous commit and
  succeeds after this change.
* We send notifications on `drop` to ensure they get sent in all cases, but this
  requires running async code in a drop handler, which is always best to avoid
* Sending the notifications as part of the database transaction greatly complicates
  retries. Retries within a transaction level don't make much sense with Postgres,
  as most errors that can happen result in either the connection being closed or
  the transaction being aborted, thus causing subsequent retries using the same
  transaction to also fail. We'd like to retry the whole sequence of
  - open transaction
  - store fetched object
  - commit

  Several times until it succeeds. However, the previous fix involved sending
  notifications out even if the transaction failed, so adding retries at the
  higher level would result in duplicate notifications being sent out. This is
  not a huge problem, but it's wasteful, unnecessary, and indicative of a
  problematic design.

All of these problems have to do with the fact that notifications are tied to
database transactions, when really they should be part of a higher level operation
that spans the entire lifetime of a transaction. On a related node, `NotifyStorage`
was adding a lot of unnecessary complexity (e.g. queuing of pending notifications)
just to abstract away notifications, and this wasn't even a good abstraction, as the
number of bugs we've seen in this area can attest.

This change does away with this abstraction entirely. `NotifyStorage` is gone, and
notifications are now handled by the `FetchingDataSource` itself, at the highest
possible level. This allows us to handle retries nicely (future work) and ensure
notifications get sent out even if we aren't able to open a transaction to store a
fetched object.

This simple idea necessitated a broader change: previously, there has been an entrypoint
to `FetchingDataSource` at the transaction level that must trigger notifications:
`UpdateAvailabilityData`. This change modifies this trait to work at the data source level
instead of the transaction level, so within it, we can handle the notifications alongside
the whole transaction life cycle, from open to write to commit. This forces us to combine
storing leaves, blocks, and VID information into a single operation, which unlocks yet
another feature that was impossible before: `FetchingDataSource` now automatically triggers
a fetch for a missing payload and/or VID data if not available at decide, rather than waiting
for the proactive scanner to discover the missing data.