Add retries with exponential backoff when storing fetched objects
Uses a crate for randomized exponential backoff. This is important
here, as one of the common causes of errors is multiple fetches
completing at the same time, leading to concurrent transactions
that can't be serialized. The randomization will cause such dueling
transactions to space themselves apart after a retry or two.

For consistency, retries in fetching itself now also use the same
exponential backoff config.