Make `update` fallible again
A recent commit made `update` infallible, reasoning that if we fail
to store a new leaf, it is better to at least try storing the next
one, rather than failing completely. However, this is not accurate:
the client uses the success/failure of `udpate` to determine when
it is safe to garbage collect data from temporary storage. Thus,
returning success when we didn't actually store all the leaves can
lead to data loss.

This change fixes the situation by again returning an error whenever
we fail to insert a leaf. It also tries to make the return value more
useful, by returning on error the height of the first leaf we failed
to insert (which is useful for garbage collection) instead of just an
opaque error message which the caller can't do much with.