Trigger fetch of missing leaf chain on leaf fetch
If we do not have a leaf and also do not have the next leaf, so
that we cannot directly fetch the required leaf, at least we can
find the first leaf that we do have and trigger a fetch, which in
turn will trigger a fetch of its parent, and so on.

This brings the behavior of leaf fetching closer to before, where
if a leaf is requested and we don't have it, we will immediately
start working to get it.