Add Inscriptions Service
The initial implementation of the Inscriptions service is
one that allows for the configuration and handling of the
various inscriptions service.

This implementation allows for the publishing of new
inscriptions with a rate limiter and a capacity in order to
try and prevent overburdening the sequencer nodes.

This implementation allows for the streaming of
new inscriptions to users that connect to the service.

Incoming submitted inscriptions are signed by the user
and verified by the service being submitted to the
mempool for sequencing.

Sequenced inscriptions are signed by the service and
verified when processing incoming blocks.