Fix uint64 underflow in SequencerClient
This underflow causes the fund-builder service to query for the block
with block number max(uint64) when the demo is started and the current
block height is still zero.

To test run `just demo` locally and `scripts/smoke-test-demo` in another
terminal.

We should also add a regression tests for this, or run our demo with
non-release build (enabled overflow checks).