Add persistence layer to the Inscriptions Service
The inscriptions service needs to be able to recover it's state when
restarting the service.  As a result, there needs to be a persistence layer
that can track where the service currently is in terms of processing
requests.

This adds a backed persistence layer to the service in terms of postgres,
with a wrapping local cache to cache the block height.