Modify block stream handling logic
The block stream has been exiting prematurely due to reasons that are not
easily trackable.  When this state is achieved, recovery does not occur, and
the service must be restarted.

In order to better understand the nature of these errors, and to improve the
exit conditions of this service, the logic for retrieving and consuming the
block stream has been adjusted.

If a block comes in that precedes the expected block, a info log will be
performed.

If the sender is determined to be closed, a panic will ensue.