Relocate pending inscription population
The pending inscription population on start could get stuck in a deadlock
due to attempting to write into a channel that is not actively being
consumed.

This change relocates where this load occurs in order to allow it to
be processed passively so that the service doesn't just stall on launch.