Avoid unnecessary payload loading.
Both the scanner and aggregator tasks wait for full blocks to be
available, but neither of them actually need the full block data.
The aggregator only needs a summary, and the scanner only needs to
know that the block exists. It is quite wasteful to load the whole
payload, especially in the scanner where we really don't care about
blocks we already have.

This change adds a new fetchable time, `PayloadMetadata`, which
summarizes a payload without including the entire payload data, and
uses it instead of the full block in both of these tasks.

Good future work would be to use this new type in many of the explorer
summary and detail queries, which currently load the whole block even
though they don't need the payload data.