Chunk aggregator task
Previously, the aggregator task fetched one block at a time and then
stored aggregate statistics for that block before moving onto the
next one. This is fine and unavoidable when we are following the
chain tip and always waiting for a new block to be produced. But this
is very inefficient and causes heavy write load on the database when
we are catching up and aggregating many old blocks.