Avoid deadlocking fetches on database failure
Previously, when there is a database failure (e.g. failure to acquire
connection from the pool) during a fetch, such that we cannot determine
whether the object exists locally or is fetchable, we would simply not
spawn an active fetch, which could result in a passive fetch that will
never resolve, even though it "should" eventually resolve, were the
database healthy.

This change addresses this issue by
* making `fetch` fallible, so that we can distinguish between three
  cases:
  - Successfully spawned a fetch
  - Successfully determined that a fetch could not be spawned
    (object is not fetchable)
  - Failed either of the above
* making `get` and `get_chunk` retry until `fetch` "succeeds" (one of
  first two cases), ensuring that we will eventually fetch every object
  that is fetchable
* these retries happen in a spawned task, so the original call returns
  quickly regardless