Modify persistance retrieval to improve performance
For each of the retrieved inscriptions we attempt to record it within the persistence
object. Retrieving the persistence object requires acquiring a read-lock on the data_state
object. With its current implementation, this read lock is acquired for each inscription,
which unnecessarily acquires the read lock in order to acquire the persistence.

This has been re-written to acquire the persistence object once, and then each inscription
is then persisted.