Use espresso-types without testing feature (#2283)
* Use espresso-types without testing feature

We would like to be able to use the types crate in other crates but
currently it won't compile without the testing feature enabled. This PR
fixes that.

- Use `NullStateCatchup` instead of `MockStateCatchup` where we don't
  need catchup.
- Feature gate `MockStateCatchup` behind testing in types crate.
- Remove the unused `InstanceState` argument from
  `from_transactions_sync`.

* Clean up features in sequencer crate

- `testing` feature enables `testing` feature of types and utils crate.
- Running tests enables `testing` feature of crate via dev-dependencies.

This allows to run `cargo check` and `cargo check --tests` in the
sequencer crate. Previously `cargo check` was broken without the
`testing` feature enabled explicitly.

* WIP: remove testing feature from non test code

* Update query service

* refactor: remove libp2p feature

The libp2p feature was always turned on and the code didn't compile
without it. With this change all feature sets checked with `cargo hack
check --each-feature` compile successfully.

* ci: check each cargo feature

* check compilation of each feature with --tests

* CI: run cargo check for feature combinations