Reconnect L1 WS stream if disconnected (#2287)
* Reconnect L1 WS stream if disconnected

Ethers internally tries to reconnect once, but if that fails, it
gives up and shuts down the whole provider backend. This is not
good in the case of extended outages of the L1 server.

This change handles errors from the underlying RPC client indicating
that the connection is closed by spawning a task to reestablish the
connection. This task tries until it succceeds.

* Cancel reconnect task if L1 client is shut down

* Rename L1 client lifecycle methods