Decrease load on DB (#2291)
* Limit number of simultaneous proposal fetches

We have seen that in some cases, when storage is slow or the node is
way behind on its event stream, it can trigger many fetches of old
proposals, which other nodes have all garbage collected. These fetches
accumulate over time, leading to a worsening problem.

This change limits the number of parallelism in proposal fetching.
Instead of spawning a new task dynamically each time we need to fetch
a proposal, we spawn a fixed number of worker tasks, each of which
will only fetch one proposal at a time. A scanner task follows the
event stream and detects when a proposal needs to be fetched, then
broadcasts the request to fetch it to the worker tasks. It will be
picked up when a worker is free.

* Decrease DB load by not storing full payload with anchor leaf

* Handle genesis payload edge case

* Avoid busy loop