Add endpoint to retrieve inscriptions for wallet address
In order to close out the inscriptions demo, the ending page requires us being able to
retrieve the list of inscriptions for the user to see his / her contribution to the
block chain.

This change adds the endpoint to retrieve a list of the most recent inscriptions that
the given wallet address made to the mainnet block chain.