Fix missing time on first block
The first block is always missing time at the moment, as the query can only
compare the time to the previous row.  Since we'll always be missing the
previous row on the first entry, we end up missing the `time` for the first
entry as well.  The fix is simply to pull one more row than we want, and to
remove the first entry when the time comes.  As a result the `Vec` has
been replaced with a `VecDeque`.

Additionally the hard-coded values have been replaced with constants with
comments.