Add `payload_size` to `ExplorerTransaction` impl
The `hotshot-query-service` has an updated requirement for the trait
`ExplorerTransaction` which needs to be implemented.  The primary
reason for this new method is to correctly reflect the actual transaction
size, as it currently reflects the size of the entire block, instead of the
size of the transaction itself.  In order for this to be adjusted the
`ExplorerTransaction` trait within the `hotshot-query-service` library
has been modified to require the `payload_size` method so the
transaction is able to specify its own size.

This change implements the new method by passing the method call
processing down to the underlying payload, which should reflect the
actual size of the transaction itself.