Jb/l1 client (#2434)
* Use HTTP for non-subscription L1 requests

There are many reasons to avoid using WebSockets for regular RPC
requests. Some of these reasons are enumerated
[here](https://docs.alchemy.com/reference/best-practices-for-using-websockets-in-web3#reasons-to-use-https-instead-of-websockets-for-json-rpc-node-requests)

This change reorganizes the L1 client so that instead of having an
HTTP provider _or_ a WebSockets provider, we have both. The HTTP
provider is used for regular RPC requests and the WebSockets provider
is used for subscriptions only. Each type of provider is now doing
what it does best, and this is evidenced in the code changes:
* we delete the complicated reconnect handling for WebSockets on failed
  RPC requests. HTTP was never subject to this complexity because each
  request is already a separate connection
* we no longer need stub methods for subscriptions for HTTP provider because
  we are no longer trying to present a unified interface that can be implemented
  by HTTP or WebSockets
* the timeout logic for subscriptions, while still in the code, is less important
  because it never affected WebSockets in the first place

We have kept the WebSockets provider as optional, since not all providers support
WebSockets. In case it is not provided, we will fall back to the old HTTP polling-based
implementation of subscriptions, which is why we need to keep the timeout logic in there.

See #2360

* L1 client failover

Allow the L1 client to be configured with multiple RPC provider URLs.
Fail over to the next provider in the list if
* two requests fail in a given time period
* N requests fail in a row

Handle rate limit errors specially by preventing requests to the
provider until the rate limit period expires.

Closes #2360

* More metrics for L1 client

* Fix panic in nasty client

* Use tokio RwLock