Replace hard-coded version constants with derived values
The `TestConfig` `new` function references two hard-coded values
related to the version of the sequencer to run.  These values have been
hard coded to `2` and `3` to indicate the versions of the sequencer that
we have been developing with.  However, the version `3` and the logic
that ensues from it is based on the assumption that version `3` will be
the marketplace, which is the value it has had historically up to this point.

Because of the change of this version's number, and the errors that
will result when we assign some other version of the sequencer to `3`,
We've opted to instead reference the static version values of the
allocated values for each of the versions for this transition.  This should
make it clear, and accurate, of which sequencer version we're
referencing, and at which point they should change.